move dumb sheet and cleanup
