--- shared/dev/dev-container.native.js
@@ -1,19 +1,0 @@
-// @flow
-import React, {Component} from 'react'
-import {Box, Text} from '../common-adapters'
-import {globalColors} from '../styles'
-
-class Container extends Component {
-  render () {
-    return (
-      <Box>
-        <Text type='Header' style={{backgroundColor: globalColors.blue, padding: 10, textAlign: 'center'}}>{this.props.title}</Text>
-        <Box style={{margin: 10, ...this.props.style}}>
-          {this.props.children}
-        </Box>
-      </Box>
-    )
-  }
-}
-
-export default Container
