+++ shared/dev/dumb-sheet/item.js.flow
@@ -1,6 +1,6 @@
 // @flow
 import {Component} from 'react'
-import type {PropsOf} from '../constants/types/more'
+import type {PropsOf} from '../../constants/types/more'
 
 export type Props<C> = {
   component: Class<C>,
